Support construction of fraction from floating-point
- only for signed integers where num and denom are same type
- deduction guides figure out workable width
- factory function, cnl::_impl::make_fraction(Float)
- algorithm from http://coliru.stacked-crooked.com/a/4f461f2ef830aa1a
  provided by Jorg Brown
- very wobbly still

+++ src/test/fraction/fraction.cpp
@@ -48,7 +48,7 @@ namespace {
                 cnl::fraction<short>(123),
                 cnl::fraction(short{123})), "cnl::fraction argument deduction");
         static_assert(identical(
-                cnl::fraction<float>(3.14285714285714285714f),
+                cnl::fraction<int>(3.14285714285714285714f),
                 cnl::fraction(3.14285714285714285714f)), "cnl::fraction argument deduction");
 
         static_assert(identical(cnl::fraction<int, int>{4321, 1}, cnl::fraction{4321}));
